Refactor: move readings into a separate module and clean up existing sensor node implementation; adjust timestamp on ble side; update naming and improve semantics;
